refactor: update import paths to use new alias structure and rename service files
