+++ src/app/provider.tsx
@@ -1,11 +1,11 @@
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
 
-interface Props {
+interface AppProviderProps {
   children: React.ReactNode;
 }
 
-export const AppProvider = ({ children }: Props) => {
+export const AppProvider = ({ children }: AppProviderProps) => {
   return (
     <QueryClientProvider client={queryClient}>
       <ReactQueryDevtools initialIsOpen={false} />
