+++ src/app/router.tsx
@@ -1,6 +1,6 @@
+import { privateRoutes, publicRoutes } from "@app/routes/index";
+import { NotFound } from "@components/common/404/not-found";
 import { createBrowserRouter } from "react-router-dom";
-import { privateRoutes, publicRoutes } from "@/app/routes";
-import { NotFound } from "@/components/common/404/not-found";
 
 export const router = createBrowserRouter([
   ...publicRoutes,
