+++ src/app/routes/index.ts
@@ -1,2 +1,2 @@
-export * from "./private.routes";
-export * from "./public.routes";
+export * from "./private-routes";
+export * from "./public-routes";
