+++ src/app/routes/public-routes.tsx
@@ -1,0 +1,15 @@
+import { AuthPage } from "@features/auth/index";
+import { PublicLayout } from "@layouts/public-layout";
+
+export const publicRoutes = [
+  {
+    path: "/",
+    element: <PublicLayout />,
+    children: [
+      {
+        path: "authenticate",
+        element: <AuthPage />
+      }
+    ]
+  }
+];
