--- src/app/routes/public.routes.tsx
@@ -1,10 +1,0 @@
-import { PublicLayout } from "@/layouts/public-layout";
-import { RouteItems } from "./route-items";
-
-export const publicRoutes = [
-  {
-    path: "/",
-    element: <PublicLayout />,
-    children: RouteItems.public
-  }
-];
