+++ src/components/layout/header.tsx
@@ -1,5 +1,5 @@
+import { Button } from "@components/ui/button";
 import { Menu } from "lucide-react";
-import { Button } from "@/components/ui/button";
 
 interface SidebarProps {
   sidebarOpen: boolean;
