+++ src/features/auth/components/register/register.tsx
@@ -1,12 +1,12 @@
+import { FormFieldWithTypes } from "@components/common/form/form-field-with-types";
+import { Button } from "@components/ui/button";
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@components/ui/card";
+import { Form } from "@components/ui/form";
+import { useRegister } from "@features/auth/hooks";
+import { registerSchema } from "@features/auth/schemas/auth.schema";
 import { zodResolver } from "@hookform/resolvers/zod";
+import type { RegisterData } from "@services/auth/auth-types";
 import { useForm } from "react-hook-form";
-import { FormFieldWithTypes } from "@/components/common/form/form-field-with-types";
-import { Button } from "@/components/ui/button";
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card";
-import { Form } from "@/components/ui/form";
-import { useRegister } from "@/features/auth/hooks";
-import { registerSchema } from "@/features/auth/schemas/auth.schema";
-import type { RegisterData } from "@/services/auth/auth.types";
 
 export const Register = () => {
   const form = useForm<RegisterData>({
