+++ src/features/restaurants/components/restaurant-list/restaurant-list.tsx
@@ -1,8 +1,8 @@
+import { RestaurantCard } from "@features/restaurants/components/restaurant-card/restaurant-card";
+import { RestaurantCardSkeleton } from "@features/restaurants/components/skeletons/restaurant-card-skeleton";
+import { useGetRestaurants } from "@features/restaurants/hooks";
+import type { RestaurantDialogData } from "@features/restaurants/types/index";
 import type { Dispatch, SetStateAction } from "react";
-import { RestaurantCard } from "@/features/restaurants/components/restaurant-card/restaurant-card";
-import { RestaurantCardSkeleton } from "@/features/restaurants/components/skeletons/restaurant-card-skeleton";
-import { useGetRestaurants } from "@/features/restaurants/hooks";
-import type { RestaurantDialogData } from "@/features/restaurants/types/index.types";
 
 interface RestaurantsProps {
   setIsDialogOpen: Dispatch<SetStateAction<RestaurantDialogData>>;
