+++ src/features/restaurants/components/restaurant/restaurant.tsx
@@ -1,7 +1,7 @@
+import { RestaurantCard } from "@features/restaurants/components/restaurant-card/restaurant-card";
+import { RestaurantCardSkeleton } from "@features/restaurants/components/skeletons/restaurant-card-skeleton";
+import { useGetRestaurant } from "@features/restaurants/hooks";
 import { useSearchParams } from "react-router-dom";
-import { RestaurantCard } from "@/features/restaurants/components/restaurant-card/restaurant-card";
-import { RestaurantCardSkeleton } from "@/features/restaurants/components/skeletons/restaurant-card-skeleton";
-import { useGetRestaurant } from "@/features/restaurants/hooks";
 
 export const Restaurant = () => {
   const [searchParams] = useSearchParams();
