+++ src/features/restaurants/components/skeletons/menu-stable-skeleton.tsx
@@ -1,4 +1,4 @@
-import { Skeleton } from "@/components/ui/skeleton";
+import { Skeleton } from "@components/ui/skeleton";
 import {
   Table,
   TableBody,
@@ -6,7 +6,7 @@ import {
   TableHead,
   TableHeader,
   TableRow
-} from "@/components/ui/table";
+} from "@components/ui/table";
 
 export const MenuTableSkeleton = () => {
   const rows = 5;
