+++ src/features/restaurants/components/skeletons/restaurant-card-skeleton.tsx
@@ -1,6 +1,6 @@
-import { Card, CardContent, CardHeader } from "@/components/ui/card";
-import { Label } from "@/components/ui/label";
-import { Skeleton } from "@/components/ui/skeleton";
+import { Card, CardContent, CardHeader } from "@components/ui/card";
+import { Label } from "@components/ui/label";
+import { Skeleton } from "@components/ui/skeleton";
 
 interface RestaurantCardSkeletonProps {
   withActionZone?: boolean;
