+++ src/features/restaurants/pages/restaurants-page.tsx
@@ -1,7 +1,7 @@
+import { RestaurantDialog } from "@features/restaurants/components/dialog/restaurant-dialog";
+import { RestaurantList } from "@features/restaurants/components/restaurant-list/restaurant-list";
+import type { RestaurantDialogData } from "@features/restaurants/types/index";
 import { useState } from "react";
-import { RestaurantDialog } from "@/features/restaurants/components/dialog/restaurant-dialog";
-import { RestaurantList } from "@/features/restaurants/components/restaurant-list/restaurant-list";
-import type { RestaurantDialogData } from "@/features/restaurants/types/index.types";
 
 export const RestaurantsPage = () => {
   const [isDialogOpen, setIsDialogOpen] = useState<RestaurantDialogData>({
