+++ src/lib/show-toast.tsx
@@ -1,6 +1,6 @@
+import { Toast } from "@components/common/toast/toast";
 import { toast } from "sonner";
-import { Toast } from "@/components/common/toast/toast";
-import type { ToastData } from "@/types/toast.types";
+import type { ToastData } from "@/types/index";
 
 export function showToast({ title, description, variant }: ToastData) {
   toast.custom(() => <Toast title={title} description={description} variant={variant} />);
