+++ src/main.tsx
@@ -1,7 +1,7 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import "./styles/global.css";
-import App from "./app/app";
+import App from "@app/app";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
