+++ src/services/auth/auth-types.ts
@@ -1,0 +1,8 @@
+export type RegisterData = {
+  firstName: string;
+  lastName: string;
+  email: string;
+  password: string;
+};
+
+export type LoginData = Pick<RegisterData, "email" | "password">;
