--- src/services/auth/auth.types.ts
@@ -1,11 +1,0 @@
-export interface RegisterData {
-  firstName: string;
-  lastName: string;
-  email: string;
-  password: string;
-}
-
-export interface LoginData {
-  email: string;
-  password: string;
-}
