+++ src/services/user/user-types.ts
@@ -1,8 +1,8 @@
-export interface UserData {
+export type UserData = {
   id: number;
   firstName: string;
   lastName: string;
   email: string;
   createdAt: string;
   updatedAt: string;
-}
+};
