+++ tsconfig.json
@@ -13,8 +13,14 @@
     "noUnusedParameters": true,
     "baseUrl": ".",
     "paths": {
-      "@/*": ["./src/*"]
+      "@/*": ["./src/*"],
+      "@app/*": ["src/app/*"],
+      "@components/*": ["src/components/*"],
+      "@features/*": ["src/features/*"],
+      "@layouts/*": ["src/layouts/*"],
+      "@lib/*": ["src/lib/*"],
+      "@services/*": ["src/services/*"],
     },
-    "types": ["vite/client"],
+    "types": ["vite/client"]
   }
 }
