+++ vite.config.ts
@@ -8,6 +8,12 @@ export default defineConfig({
   resolve: {
     alias: {
       "@": path.resolve(__dirname, "./src"),
+      "@app": path.resolve(__dirname, "src/app"),
+      "@components": path.resolve(__dirname, "src/components"),
+      "@features": path.resolve(__dirname, "src/features"),
+      "@layouts": path.resolve(__dirname, "src/layouts"),
+      "@lib": path.resolve(__dirname, "src/lib"),
+      "@services": path.resolve(__dirname, "src/services"),
     },
   },
 });
